Fix #24

+++ app/helpers/projects_helper.rb
@@ -1,14 +1,18 @@
 module ProjectsHelper
 
   def heart_link
-    if @liked
-      link_to unlike_project_path(@project), method: :delete do
-        '<i class="fa fa-heart text-danger"></i>'.html_safe
+    if can? :create, ProjectLike
+      if @liked
+        link_to unlike_project_path(@project), method: :delete do
+          '<i class="fa fa-heart text-danger"></i>'.html_safe
+        end
+      else
+        link_to like_project_path(@project), method: :post do
+          '<i class="fa fa-heart-o pull-right"></i>'.html_safe
+        end
       end
     else
-      link_to like_project_path(@project), method: :post do
-        '<i class="fa fa-heart-o pull-right"></i>'.html_safe
-      end
+      '<i class="fa fa-heart-o pull-right"></i>'.html_safe
     end
   end
